// models/character.model.ts
export interface CharacterModel {
    name: string;
    height: string;
    mass: string;
    hair_color: string;
    skin_color: string;
    eye_color: string;
    birth_year: string;
    gender: string;
    homeworld: string;
    films: string[];
    species: string[];
    vehicles: string[];
    starships: string[];
    created: string;
    edited: string;
    url: string;
}

export interface CharacterSpanishModel {
    nombre: string;
    altura: string;
    masa: string;
    color_cabello: string;
    color_piel: string;
    color_ojos: string;
    ano_nacimiento: string;
    genero: string;
    planeta_natal: string;
    peliculas: string[];
    especies: string[];
    vehiculos: string[];
    naves_estelares: string[];
    creado: string;
    editado: string;
    url: string;
}